% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{at_transform}
\alias{at_transform}
\title{Transform data}
\usage{
at_transform(mod, df, var_name, value)
}
\arguments{
\item{df}{dataframe}

\item{var_name}{variable name}

\item{value}{value of variable}
}
\value{
dataframe with transformed variable
}
\description{
The reason why you want to use this instead of ?transform is that
the syntax for factors is really annoying.
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mod <- glm(cyl ~ gear + mpg, data = mtcars)
df3 <- transform(mtcars[, c('cyl', 'gear', 'mpg')],
                 gear = factor(3, levels = levels(mtcars$gear)))
df <- at_transform(mod = mod, df = mtcars, var_name = 'gear', value = 3)
str(df3)
str(df)
all(df == df3)
}

