% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{at_transforms}
\alias{at_transforms}
\title{Apply multiple transformations}
\usage{
at_transforms(df, var_name, values)
}
\arguments{
\item{df}{dataframe}

\item{var_name}{variable}

\item{values}{vector of values for that variable}
}
\value{
list of transformed dataframes
}
\description{
Apply multiple transformations
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
df3 <- transform(mtcars, gear = factor(3, levels = levels(mtcars$gear)))
df <- at_transforms(df = mtcars, var_name = 'gear', values = c(3, 5))
}

