% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_jacob.R
\name{calc_jacob}
\alias{calc_jacob}
\title{Calculate the Jacobian matrix for predictive levels or effects}
\usage{
calc_jacob(pred_values, covar_matrix, deriv_func)
}
\arguments{
\item{pred_values}{numeric vector of predicted values}

\item{covar_matrix}{numeric matrix of covariates}

\item{deriv_func}{if levels, this should be the function for the derivative
of the link function with respect to eta. if effects, this should be
the second derivative of the above.}
}
\value{
numeric vector row of jacobian
}
\description{
Calculate the Jacobian matrix for predictive levels or effects
}
\examples{

data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear + mpg * disp, mtcars, family = 'binomial')

new_df <- at_transform(df = mtcars, var_name = 'gear', value = 3)

calc_jacob(
  pred_values = predict(
    mm, new_df),
  covar_matrix = model.matrix(
    object = mm$formula, data = new_df, contrasts.arg = mm$contrasts,
    xlev = mm$xlevels)[, !is.na(coef(mm))],
  deriv_func = mm$family$mu.eta)

}

