% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pred_se.R
\name{calc_pred_se}
\alias{calc_pred_se}
\title{Calculate margins and se}
\usage{
calc_pred_se(df_trans, var_interest, model)
}
\arguments{
\item{df_trans}{data.frame, should already be transformed for variables not related to
the variable of interest}

\item{var_interest}{the variable of interest}

\item{model}{model}
}
\value{
dataframe of formatted output
}
\description{
For one set of transformed covariates (not including the variable of interest),
calculate the predicted level and se for the variable of interest.
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear + mpg * disp, mtcars, family = 'binomial')
# apply at transformations
df <- at_transforms(mm$model, list("mpg" = c(15, 21)))
df <- df[[1]]
calc_jac_se(df, var_interest = 'gear', model = mm)
}

