% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_effects.R
\name{discrete_effect_jacob}
\alias{discrete_effect_jacob}
\title{Calculate jacobian for effects}
\usage{
discrete_effect_jacob(jacobian, base_rn = 1)
}
\arguments{
\item{jacobian}{numeric matrix of jacobian for all levels}

\item{base_rn}{row number of the base level, defaults to 1}
}
\value{
numeric matrix, jacobian of difference
}
\description{
Calculate jacobian for effects
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear + mpg * disp, mtcars, family = 'binomial')

new_dfs <- lapply(3:5, function(x){
  at_transform(df = mtcars, var_name = 'gear', value = x)})
z <- do.call(
  rbind, lapply(new_dfs, function(x){
    calc_jacob(
      pred_values = predict(mm, x),
      covar_matrix = model.matrix(
        object = mm$formula, data = x, contrasts.arg = mm$contrasts,
        xlev = mm$xlevels)[, !is.na(coef(mm))],
      deriv_func = mm$family$mu.eta)
  }))

discrete_effect_jacob(z, 1)

}

