% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_effects.R
\name{discrete_effect_pred}
\alias{discrete_effect_pred}
\title{Calculate predictions for effects}
\usage{
discrete_effect_pred(pred, base_rn = 1)
}
\arguments{
\item{pred}{numeric vector of predictions for all levels}

\item{base_rn}{row number of the base level, defaults to 1}
}
\value{
numeric vector, difference of predictions
}
\description{
Calculate predictions for effects
}
\examples{
discrete_effect_pred(c(1,3),1)
}

