% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_wrap.R
\name{discrete_wrap}
\alias{discrete_wrap}
\title{Calculate margins and se for discrete variables}
\usage{
discrete_wrap(df_trans, var_interest, model, type = "levels", base_rn = 1,
  at_var_interest = NULL, vcov_mat = NULL)
}
\arguments{
\item{df_trans}{data.frame, should already be transformed for variables not related to
the variable of interest}

\item{var_interest}{the variable of interest}

\item{model}{model}

\item{type}{either effects or levels, defaults to levels}

\item{base_rn}{row number of the base level, defaults to 1}

\item{vcov_mat}{variance-covariance matrix, defaults to NULL}
}
\value{
dataframe of formatted output
}
\description{
For one set of transformed covariates (not including the variable of interest),
calculate the predicted level and se for the variable of interest.
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear + mpg * disp, mtcars, family = 'binomial')
# apply at transformations
df <- at_transforms(mm$model, list("mpg" = c(15, 21)))
df <- df[[1]]
discrete_wrap(df, var_interest = 'gear', model = mm)
}

