% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_margins.R
\name{get_margins}
\alias{get_margins}
\title{Get Margins}
\usage{
get_margins(preds)
}
\arguments{
\item{preds}{list of individual predictive margins}
}
\value{
vector of means
}
\description{
Get Margins
}
\examples{
get_margins(list(1:3, 8:10))
}

