% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacob_discrete.R
\name{jacob_discrete}
\alias{jacob_discrete}
\title{Jacobian matrix to discrete}
\usage{
jacob_discrete(pred_values, covar_matrix, link_deriv, base_rn = 1,
  effects = F)
}
\arguments{
\item{pred_values}{numeric vector of predicted values}

\item{covar_matrix}{numeric matrix of covariates}

\item{link_deriv}{function for the derivative of the link function with respect to eta}

\item{base_rn}{row number of the base level, defaults to 1}
}
\value{
jacobian matrix
}
\description{
Jacobian matrix to discrete
}
\examples{
jacob_discrete(pred_values = z$pred, covar_matrix = z$covar,
link_deriv = mm$family$mu.eta, effects = T)
}

