% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacob_level.R
\name{jacob_level}
\alias{jacob_level}
\title{Calculate the Jacobian matrix for predictive levels}
\usage{
jacob_level(pred_values, covar_matrix, link_deriv)
}
\arguments{
\item{pred_values}{numeric vector of predicted values}

\item{covar_matrix}{numeric matrix of covariates}

\item{link_deriv}{function for the derivative of the link function with respect to eta}
}
\value{
numeric vector row of jacobian
}
\description{
Calculate the Jacobian matrix for predictive levels
}
\examples{

data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear + mpg * disp, mtcars, family = 'binomial')

binom_family <- make.link('logit')
ld_fun <- binom_family$mu.eta

df3 <- transform(mm$data, gear = factor(3, levels = levels(mtcars$gear)))
df4 <- transform(mm$data, gear = factor(4, levels = levels(mtcars$gear)))
df5 <- transform(mm$data, gear = factor(5, levels = levels(mtcars$gear)))

covar3 <- model.matrix(mm$formula, df3)
covar4 <- model.matrix(mm$formula, df4)
covar5 <- model.matrix(mm$formula, df5)

# Note type != response
p3 <- predict(mm, newdata = df3)
p4 <- predict(mm, newdata = df4)
p5 <- predict(mm, newdata = df5)

jacob_level(p3, as.matrix(covar3), ld_fun)
jacob_level(p4, as.matrix(covar4), ld_fun)
jacob_level(p5, as.matrix(covar5), ld_fun)

}

