% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrobust_marg.R
\name{marg.lm_robust}
\alias{marg.lm_robust}
\title{Predicted Margins for `estimatr::lm_robust` objects}
\usage{
\method{marg}{lm_robust}(mod, var_interest, data, weights = mod$weights, ...)
}
\arguments{
\item{mod}{model object, currently only support those of class \code{\link[stats]{glm}}
or \code{\link[AER]{ivreg}}}

\item{var_interest}{name of the variable of interest, must correspond to a
covariate in the model}

\item{data}{data.frame that margins should run over, defaults changes based
on class-specific method}

\item{weights}{numeric, vector of weights used to generate predicted levels,
defaults changes based on class-specific method. Must be equal to the number
of rows in \code{data}.}

\item{...}{additional parameters passed to \code{?marg}.}
}
\description{
Obtains predicted margins and standard errors
of those predictions from a fitted generalized linear model object.
}
\examples{

}
