% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_effects.R
\name{pred_effect}
\alias{pred_effect}
\title{Calculate predictions for effects}
\usage{
pred_effect(pred, base_rn = 1)
}
\arguments{
\item{pred}{numeric vector of predictions for all levels}

\item{base_rn}{row number of the base level, defaults to 1}
}
\value{
numeric vector, difference of predictions
}
\description{
Calculate predictions for effects
}
\examples{
pred_effect(c(1,3),1)
}

