% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_se.R
\name{pred_se}
\alias{pred_se}
\title{Calculate margins and se}
\usage{
pred_se(df_trans, var_interest, model, type = "levels", base_rn = 1,
  at_var_interest = NULL, vcov_mat = NULL)
}
\arguments{
\item{df_trans}{data.frame, should already be transformed for variables not related to
the variable of interest}

\item{var_interest}{the variable of interest}

\item{model}{model}

\item{type}{either effects or levels, defaults to levels}

\item{base_rn}{row number of the base level, defaults to 1}

\item{at_var_interest}{vector, if type == 'levels', the values for the variable of interest at which levels should be calculated. if NULL, indicates all levels for a
factor variable, defaults to NULL}

\item{vcov_mat}{variance-covariance matrix, defaults to NULL}
}
\value{
list of labels, predicted margins, and SE
}
\description{
For one set of transformed covariates (not including the variable of interest),
calculate the predicted level and se for the variable of interest.
}

