% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_se_wrap.R
\name{pred_se_wrap}
\alias{pred_se_wrap}
\title{Calculate margins and se}
\usage{
pred_se_wrap(df_trans, var_interest, model, type = "levels", base_rn = 1,
  at_var_interest = NULL, vcov_mat = NULL, dof = model$df.residual)
}
\arguments{
\item{df_trans}{data.frame, should already be transformed for variables not related to
the variable of interest}

\item{var_interest}{the variable of interest}

\item{model}{model}

\item{type}{either effects or levels, defaults to levels}

\item{base_rn}{row number of the base level, defaults to 1}

\item{at_var_interest}{vector, if type == 'levels', the values for the variable of interest at which levels should be calculated. if NULL, indicates all levels for a
factor variable, defaults to NULL}

\item{vcov_mat}{variance-covariance matrix, defaults to NULL}

\item{dof}{integer degrees of freedom for T statistics}
}
\value{
dataframe of formatted output
}
\description{
For one set of transformed covariates (not including the variable of interest),
calculate the predicted level and se for the variable of interest.
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear  + mpg * disp, mtcars, family = 'binomial')
df <- transform(mm$model, mpg = 15)
pred_se_wrap(df, var_interest = 'gear', model = mm, vcov_mat = vcov(mm))
pred_se_wrap(mm$model, var_interest = 'mpg',
                at_var_interest = c(15, 21), model = mm,
                vcov_mat = vcov(mm))
}

