% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_modelmat.R
\name{predict_modelmat}
\alias{predict_modelmat}
\title{Generate predictions and model matrix}
\usage{
predict_modelmat(model, transformed_df, formula = model$formula)
}
\arguments{
\item{model}{model object}

\item{transformed_df}{df after applying whatever transform you want}

\item{formula}{model formula, defaults to model$formula}
}
\value{
list of covariates (covar), predictions on the link function scale (pred_link),
and predictions on the response variable scale (pred_response), e.g., probabilities
}
\description{
Generate predictions and model matrix
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear + mpg * disp, mtcars, family = 'binomial')

df3 <- at_transform(df = mtcars, var_name = 'gear', value = 3)
covar3 <- model.matrix(mm$formula, df3)

z <- predict_modelmat(model = mm, transformed_df = df3)
all(z$covar == covar3)
all(z$pred_link == predict(mm, newdata = df3))
all(z$pred_resp == predict(mm, newdata = df3, type = 'response'))
}

