% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvcov.R
\docType{data}
\name{rvcov}
\alias{rvcov}
\title{Robust variance-covariance matrices and T statistic d.o.f.}
\format{A list of three variance-covariance matrices, from an OLS model, logit model, and OLS with a polynomial
interaction with missing data.}
\usage{
rvcov
}
\description{
Variance-covariance matrices with robust standard errors and
degrees-of-freedom for T statistics, for tests and examples specifying \code{vcov}.
Generated with \code{margex} data in this package.
}
\details{
See data-raw/make_robust_vcov.R for details.
}
\keyword{datasets}
