% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{at_transforms}
\alias{at_transforms}
\title{Apply multiple transformations}
\usage{
at_transforms(model_df, at_list, mod)
}
\arguments{
\item{model_df}{dataframe used in model (not model.matrix)}

\item{at_list}{list of transformations, in the format of list("variable" = c("values"))}

\item{model}{model used to generate model_df}
}
\value{
list of dataframes, each transformed
}
\description{
Apply multiple transformations
}
\examples{
data(mtcars)
at_list <- list("mpg" = c(15, 21), "disp" = c(140, 180))
at_transforms(mtcars, at_list)
}

