% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pred_se.R
\name{calc_pred_se}
\alias{calc_pred_se}
\title{Calculate standard errors for predictive margins}
\usage{
calc_pred_se(vcov_model, jac)
}
\arguments{
\item{vcov_model}{variance-covariance matrix of the model}

\item{jacobian}{jacobian matrix of the predictions}
}
\value{
vector of standard errors
}
\description{
Calculate standard errors for predictive margins
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear + mpg * disp, mtcars, family = 'binomial')

binom_family <- make.link('logit')
ld_fun <- binom_family$mu.eta

df3 <- transform(mm$data, gear = factor(3, levels = levels(mtcars$gear)))
covar3 <- model.matrix(mm$formula, df3)
p3 <- predict(mm, newdata = df3)

z <-jacob_level(p3, as.matrix(covar3), ld_fun)

calc_pred_se(vcov(mm), z)
calc_pred_se(vcov(mm), rbind(z, z))
}

