% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_wrap.R
\name{continuous_wrap}
\alias{continuous_wrap}
\title{Calculate margins and se for continuous variables}
\usage{
continuous_wrap(df_trans, var_interest, model, at_var_interest = NULL,
  type = "levels", vcov_mat = NULL)
}
\arguments{
\item{df_trans}{data.frame, should already be transformed for variables not related to
the variable of interest}

\item{var_interest}{the variable of interest}

\item{model}{model}

\item{at_var_interest}{levels of variable of interest}

\item{type}{either effects or levels, defaults to levels}

\item{vcov_mat}{variance-covariance matrix, defaults to NULL}
}
\value{
dataframe of formatted output
}
\description{
For one set of transformed covariates (not including the variable of interest),
calculate the predicted level and se for the variable of interest.
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear + mpg * disp, mtcars, family = 'binomial')
# apply at transformations
df <- at_transforms(mm$model, list("mpg" = c(15, 21)))
df <- df[[1]]
continuous_wrap(df, var_interest = 'gear', model = mm, vcov_mat = vcov(mm))
}

