% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{format_output}
\alias{format_output}
\title{Format margins output}
\usage{
format_output(margin_labels, pred_margins, se, family, dof, cofint = c(0.025,
  0.975))
}
\arguments{
\item{margin_labels}{vector, label of what the margins mean}

\item{pred_margins}{vector, predictive margins}

\item{se}{vector, standard errors}

\item{family}{character, type of model (e.g. gaussian, binomial)}

\item{dof}{integer, residual degrees of freedom}

\item{cofint}{vector, confidence intervals defaults to c(0.025, 0.975)}
}
\value{
dataframe of label, margin, se, z-value, p-value, and confidence interval bounds
}
\description{
Format margins output
}
\examples{
format_output(margin_labels = c('hello', 'goodbye', 'whatever'),
pred_margins = c(1, 0.25, 3.1),
se = c(0.25, 0.75, 0.5), family = 'binomial')
}

