% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacob_level.R
\name{jacob_level}
\alias{jacob_level}
\title{Calculate the Jacobian matrix for predictive levels}
\usage{
jacob_level(pred_values, covar_matrix, inv_link_deriv)
}
\arguments{
\item{pred_values}{numeric vector of predicted values}

\item{covar_matrix}{numeric matrix of covariates}

\item{inv_link_deriv}{function for the derivative of the inverse link function with respect to eta}
}
\value{
numeric vector row of jacobian
}
\description{
Calculate the Jacobian matrix for predictive levels
}
\examples{

# Run model
data(mtcars)
mtcars$carb <- factor(mtcars$carb)
mm <- glm(vs ~ carb + mpg * disp, mtcars, family = 'binomial')

# Create transformed covariates
covar <- transform(mm$model, carb = 1)
covar$`mpg:disp` <- covar$mpg * covar$disp

# Generate predictions
z <- predict(mm, data = covar)

# Specify derivative of link function
binom_family <- make.link('logit')
ld_fun <- binom_family$mu.eta

jacob_level(pred_values = z, covar_matrix = covar, inv_link_deriv = ld_fun)
}

