% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacob_level.R
\name{jacob_level}
\alias{jacob_level}
\title{Calculate the Jacobian matrix for predictive levels}
\usage{
jacob_level(pred_values, covar_matrix, link_deriv)
}
\arguments{
\item{pred_values}{numeric vector of predicted values}

\item{covar_matrix}{numeric matrix of covariates}

\item{link_deriv}{function for the derivative of the link function with respect to eta}
}
\value{
numeric vector row of jacobian
}
\description{
Calculate the Jacobian matrix for predictive levels
}
\examples{

data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mm <- glm(vs ~ gear + mpg * disp, mtcars, family = 'binomial')

binom_family <- make.link('logit')
ld_fun <- binom_family$mu.eta

x3 <- predict_modelmat(model = mm,
transformed_df = apply_transform(df = mtcars, var_name = 'gear', value = 3)
)

x4 <- predict_modelmat(model = mm,
transformed_df = apply_transform(df = mtcars, var_name = 'gear', value = 4)
)

x5 <- predict_modelmat(model = mm,
transformed_df = apply_transform(df = mtcars, var_name = 'gear', value = 5)
)

jacob_level(x3$pred, x3$covar, ld_fun)
jacob_level(x4$pred, x4$covar, ld_fun)
jacob_level(x5$pred, x5$covar, ld_fun)

}

