% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_marg.R
\name{mod_marg2}
\alias{mod_marg2}
\title{Estimating predictive margins on a model}
\usage{
mod_marg2(mod, var_interest, type = "levels", vcov_mat = vcov(mod),
  at = NULL, base_rn = 1, at_var_interest = NULL)
}
\arguments{
\item{mod}{model object, currently only support those of class glm}

\item{var_interest}{variable of interest}

\item{type}{either 'levels' (predicted outcomes) or 'effects' (dydx), defaults to 'levels'}

\item{vcov_mat}{the variance-covariance matrix, defaults to vcov(model)}

\item{at}{list, should be in the format of list('var_name' = c(values)), defaults to NULL.
This calculates the margins of the variable at these particular variables.}

\item{base_rn}{numeric, if type == 'effects', the base level (taken as the index of one of
the ordered unique values in var_interest). if type == 'levels', this param is ignored.
Defaults to 1.}

\item{at_var_interest}{vector, if type == 'levels', the values for the variable of interest at which levels should be calculated. if NULL, indicates all levels for a
factor variable, defaults to NULL}
}
\value{
list of dataframes with predicted margins/effects, se, p-values, and confidence interval bounds
}
\description{
Estimating predictive margins on a model
}
\details{
P values are calculated with T tests for OLS, and Z tests otherwise.
}
\examples{
data(mtcars)
mtcars$gear <- factor(mtcars$gear)
mod <- glm(vs ~ gear + mpg * disp, data = mtcars, family = 'binomial')
mod_marg2(mod, var_interest = 'gear',
          type = 'levels', at = list(mpg = c(15, 21), disp = c(140, 180)))

data(margex)
margex$treatment <- factor(margex$treatment)
mod <- glm(outcome ~ treatment + distance, data = margex, family = 'binomial')
mod_marg2(mod, var_interest = 'treatment', type = 'levels', at = NULL)
mod_marg2(mod, var_interest = 'treatment', type = 'effects', at = NULL)
mod_marg2(mod, var_interest = 'distance', type = 'levels',
          at = NULL, at_var_interest = c(10, 20, 30))
}

