% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_se.R
\name{pred_se}
\alias{pred_se}
\title{Main wrapper function to calculate margins and standard errors}
\usage{
pred_se(
  df_levels,
  model,
  type,
  base_rn,
  vcov_mat,
  weights,
  deriv_func,
  link_func
)
}
\arguments{
\item{df_levels}{data.frame, already transformed for variables not related
to the variable of interest}

\item{model}{model object}

\item{type}{either effects or levels}

\item{base_rn}{numeric, row number of the base level}

\item{vcov_mat}{matrix, variance-covariance matrix}

\item{weights}{vector of weights, or NULL}

\item{deriv_func}{function for the derivative of the predicted outcomes}

\item{link_func}{function to transform output of `predict` method into
response scale}
}
\description{
For one set of transformed covariates (not including the variable of
interest), calculate the predicted level and standard error for the
variable of interest.
}
